Added Hue::setColorXy and Hue::setColorRgb

+++ include/Hue/State.h
@@ -12,8 +12,8 @@ class State {
         int hue;
         int sat;
         int ct;
-        float colorX;
-        float colorY;
+        double colorX;
+        double colorY;
         bool reachable;
         char * colormode;
         static State * fromJson(cJSON * json);
